use javascript for cli
