+++ playground/sk8-ministries/package.json
@@ -9,11 +9,12 @@
     "build": "astro build",
     "preview": "astro preview",
     "astro": "astro",
-    "lightnet-cli": "node --experimental-strip-types ../../packages/cli/src/index.ts"
+    "lightnet": "lightnet"
   },
   "dependencies": {
     "@astrojs/react": "^5.0.4",
     "@lightnet/sveltia-admin": "workspace:^",
+    "@lightnet/cli": "workspace:^",
     "astro": "^6.2.2",
     "lightnet": "workspace:^",
     "react": "^19.2.5",
